add back app workers to app class
was unnecessary
