+++ dyno_viewer/app.py
@@ -3,13 +3,13 @@ from textual.widgets import (
     Footer,
 )
 from textual.reactive import reactive
-from dyno_viewer.app_workers import (
-    update_dyn_table_info,
-    dyn_table_query,
-    UpdateDynDataTable,
-)
 from dyno_viewer.aws.session import get_available_profiles
-from dyno_viewer.aws.ddb import get_ddb_client, table_client_exist
+from dyno_viewer.aws.ddb import (
+    get_ddb_client,
+    query_items,
+    scan_items,
+    table_client_exist,
+)
 from dyno_viewer.components.screens import (
     ProfileSelectScreen,
     RegionSelectScreen,
@@ -23,9 +23,18 @@ from textual import work, log, on
 import pyclip
 from itertools import cycle
 from dyno_viewer.components.types import TableInfo
+from textual.message import Message
 from dyno_viewer.util import output_to_csv_str
 
 cursors = cycle(["column", "row", "cell"])
+
+
+class UpdateDynDataTable(Message):
+    def __init__(self, data, next_token, update_existing_data=False) -> None:
+        self.data = data
+        self.next_token = next_token
+        self.update_existing_data = update_existing_data
+        super().__init__()
 
 
 class DynCli(App):
@@ -86,6 +95,62 @@ class DynCli(App):
         else:
             self.dyn_query_params.pop("ExclusiveStartKey", None)
 
+    # worker methods
+
+    @work(exclusive=True, group="update_dyn_table_info", thread=True)
+    def update_dyn_table_info(self) -> None:
+        worker = get_current_worker()
+        if not worker.is_cancelled:
+            # temp disable logging doesn't work
+            # self.log(f"updating table info")
+            # self.log("key schema=", self.table_client.key_schema)
+            # self.log("gsi schema=", self.table_client.global_secondary_indexes)
+            main_keys = {
+                ("primaryKey" if key["KeyType"] == "HASH" else "sortKey"): key[
+                    "AttributeName"
+                ]
+                for key in self.table_client.key_schema
+            }
+
+            gsi_keys = {
+                gsi["IndexName"]: {
+                    ("primaryKey" if key["KeyType"] == "HASH" else "sortKey"): key[
+                        "AttributeName"
+                    ]
+                    for key in gsi["KeySchema"]
+                }
+                for gsi in self.table_client.global_secondary_indexes or []
+            }
+
+            def update(self, main_keys, gsi_keys):
+                self.table_info = {"keySchema": main_keys, "gsi": gsi_keys}
+
+            self.call_from_thread(update, self, main_keys, gsi_keys)
+
+    @work(exclusive=True, group="dyn_table_query", thread=True)
+    def dyn_table_query(self, dyn_query_params, update_existing=False):
+        worker = get_current_worker()
+        if not worker.is_cancelled:
+            # temp disable logging doesn't work
+            # self.log("dyn_params=", app.dyn_query_params)
+            result, next_token = (
+                query_items(
+                    self.table_client,
+                    paginate=False,
+                    Limit=50,
+                    **dyn_query_params,
+                )
+                if "KeyConditionExpression" in dyn_query_params
+                else scan_items(
+                    self.table_client,
+                    paginate=False,
+                    Limit=50,
+                    **dyn_query_params,
+                )
+            )
+
+            self.post_message(UpdateDynDataTable(result, next_token, update_existing))
+
     # on methods
 
     def on_mount(self):
@@ -100,7 +165,7 @@ class DynCli(App):
             if "ExclusiveStartKey" in self.dyn_query_params:
                 log.info("adding more items")
 
-                dyn_table_query(self, self.dyn_query_params, update_existing=True)
+                self.dyn_table_query(self.dyn_query_params, update_existing=True)
 
     async def on_region_select_screen_region_selected(
         self, selected_region: RegionSelectScreen.RegionSelected
@@ -137,7 +202,7 @@ class DynCli(App):
         if run_query.filter_cond_exp:
             params["FilterExpression"] = run_query.filter_cond_exp
         self.dyn_query_params = params
-        dyn_table_query(self, params)
+        self.dyn_table_query(params)
 
     async def on_update_dyn_data_table(self, update_data: UpdateDynDataTable) -> None:
         table = self.query_one(DataDynTable)
@@ -187,8 +252,8 @@ class DynCli(App):
         """update DynTable with new table data"""
         if new_table_client:
             log.info("table client changed and table found, Update table data")
-            update_dyn_table_info(self)
-            dyn_table_query(self, self.dyn_query_params)
+            self.update_dyn_table_info()
+            self.dyn_table_query(self.dyn_query_params)
         else:
             log.info("table client changed and table not found, Clear table data")
             self.query_one(DataDynTable).clear()
